llm-assisted-router/requirements.txt
aws-cdk-lib==2.162.0
constructs==10.4.0